version 1.1

+++ src/main/java/ru/yandex/practicum/tarasov/wallettesttask/DTO/BalanceDto.java
@@ -1,6 +1,7 @@
 package ru.yandex.practicum.tarasov.wallettesttask.DTO;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record BalanceDto(UUID id, long amount, String currencyCode) {
+public record BalanceDto(UUID id, BigDecimal amount, String currencyCode) {
 }
